feat(ui): Использование sealed interface вместо class для Response
