--- composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/core/data/network/dto/CheckResponse.kt
@@ -1,10 +1,0 @@
-package io.dimasla4ee.shoppinglist.core.data.network.dto
-
-import kotlinx.serialization.Serializable
-
-/** Используется в `/auth/check` эндпоинте. */
-@Serializable
-data class CheckResponse(
-    val success: Boolean,
-    val refresh: Boolean
-): Response()
