+++ composeApp/src/commonMain/kotlin/io/dimasla4ee/shoppinglist/core/data/network/dto/Response.kt
@@ -3,11 +3,45 @@ package io.dimasla4ee.shoppinglist.core.data.network.dto
 import kotlinx.serialization.Serializable
 
 /**
- * Базовый класс сетевых ответов.
+ * Базовый интерфейс сетевых ответов.
  *
- * @param resultCode HTTP-код ответа.
+ * [Документация к API](https://practicumopbackend-production.up.railway.app/swagger-ui/index.html)
  */
 @Serializable
-open class Response {
-    var resultCode = 0
+sealed interface Response {
+    /** Используется в `/auth/registration` эндпоинте. */
+    @Serializable
+    data class RegisterResponse(
+        val userId: Int,
+        val accessToken: String,
+        val refreshToken: String
+    ) : Response
+
+    /** Используется в `/auth/login` эндпоинте. */
+    @Serializable
+    data class UserAuthResponse(
+        val userId: Int,
+        val accessToken: String,
+        val refreshToken: String
+    ): Response
+
+    /** Используется в `/auth/recovery` эндпоинте. */
+    @Serializable
+    data class RecoverPasswordResponse(
+        val message: String
+    ) : Response
+
+    /** Используется в `/auth/check` эндпоинте. */
+    @Serializable
+    data class CheckResponse(
+        val success: Boolean,
+        val refresh: Boolean
+    ): Response
+
+    /** Используется в `/auth/refresh` эндпоинте. */
+    @Serializable
+    data class RefreshTokenResponse(
+        val accessToken: String,
+        val refreshToken: String
+    ) : Response
 }
